Fix : "오류 응답 시 명확한 상태코드를 반환하도록 처리" 1. 오류 응답 시 명확한 상태코드를 반환하도록 처리

+++ src/main/java/onthelive/kr/authServer/controller/ClientRegistrationController.java
@@ -2,9 +2,7 @@ package onthelive.kr.authServer.controller;
 
 import lombok.RequiredArgsConstructor;
 import onthelive.kr.authServer.entity.ClientEntity;
-import onthelive.kr.authServer.model.Client;
 import onthelive.kr.authServer.model.Registration;
-import onthelive.kr.authServer.service.AuthorService;
 import onthelive.kr.authServer.service.ClientService;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.http.HttpStatus;
@@ -12,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.Arrays;
 import java.util.Date;
 
 @RestController
@@ -20,7 +17,6 @@ import java.util.Date;
 public class ClientRegistrationController {
 
     private final ClientService clientService;
-    private final AuthorService authorService;
 
     @PostMapping("/register")
     public ResponseEntity postRegister(HttpServletRequest request) {
@@ -33,9 +29,9 @@ public class ClientRegistrationController {
         register.setResponse_types(request.getParameter("response_types"));
 
         ResponseEntity<ClientEntity> checkClientMetadata = clientService.checkClientMetadata(register);
-        ClientEntity client = checkClientMetadata.getBody();
 
         if (checkClientMetadata.hasBody() && checkClientMetadata.getStatusCodeValue() == 200) {
+            ClientEntity client = checkClientMetadata.getBody();
             client.setClientId(RandomStringUtils.randomAlphanumeric(8));
             client.setClientSecret(RandomStringUtils.randomAlphanumeric(8));
             client.setClientIdCreatedAt(new Date().getTime());
@@ -60,7 +56,6 @@ public class ClientRegistrationController {
 
             return new ResponseEntity(registration, HttpStatus.CREATED);
         }
-
         return new ResponseEntity(checkClientMetadata.getStatusCode());
     }
 
@@ -70,8 +65,8 @@ public class ClientRegistrationController {
         ResponseEntity<ClientEntity> responseEntity = clientService.authorizeConfigurationEndpointRequest(request, clientId);
         ClientEntity clientEntity = responseEntity.getBody();
 
-        if(clientEntity == null){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        if(responseEntity.getStatusCode().isError()){
+            return new ResponseEntity<>(responseEntity.getStatusCode());
         }
 
         Registration registration = new Registration(
@@ -95,20 +90,19 @@ public class ClientRegistrationController {
                             @PathVariable(name = "clientId") String clientId){
         ResponseEntity<ClientEntity> responseEntity = clientService.authorizeConfigurationEndpointRequest(request, clientId);
         ClientEntity clientEntity = responseEntity.getBody();
-        System.out.println("clientEntity = " + clientEntity);
+
+        if(responseEntity.getStatusCode().isError()){
+            return new ResponseEntity<>(responseEntity.getStatusCode());
+        }
 
         if(!request.getParameter("client_id").equals(clientEntity.getClientId())){
             return new ResponseEntity(HttpStatus.BAD_REQUEST);
         }
 
         if(request.getParameter("client_secret") == null
-                && !request.getParameter("client_secret").equals(clientEntity.getClientSecret())){
+                || !request.getParameter("client_secret").equals(clientEntity.getClientSecret())){
             return new ResponseEntity(HttpStatus.BAD_REQUEST);
         }
-
-
-        // 현재 clientEntity 는 clientId에 해당하는 모든 정보를 가지고 있다.
-        // request 에는 clientEntity의 정보 중 변경하고자 하는 정보를 가지고 있다.
 
         Registration registration = new Registration();
         registration.setGrant_types(request.getParameter("grant_types"));
@@ -116,14 +110,13 @@ public class ClientRegistrationController {
         registration.setRedirect_uris(request.getParameter("redirect_uris"));
         registration.setScope(request.getParameter("scope"));
 
-        System.out.println("registration = " + registration);
 
         ResponseEntity<ClientEntity> checkClientMetadata = clientService.checkClientMetadata(registration);
-        // TODO status code check !
-        System.out.println("checkClientMetadata.getStatusCodeValue() = " + checkClientMetadata.getStatusCodeValue());
         ClientEntity checkedClientMetaData = checkClientMetadata.getBody();
 
-        System.out.println("checkedClientMetaData = " + checkedClientMetaData);
+        if(checkClientMetadata.getStatusCode().isError()){
+            return new ResponseEntity<>(checkClientMetadata.getStatusCode());
+        }
 
         Registration modifyResult = clientService.modifyClient(checkedClientMetaData, clientEntity);
 
@@ -139,7 +132,9 @@ public class ClientRegistrationController {
                                @PathVariable(name = "clientId") String clientId){
         ResponseEntity<ClientEntity> responseEntity = clientService.authorizeConfigurationEndpointRequest(request, clientId);
         ClientEntity clientEntity = responseEntity.getBody();
-
+        if(responseEntity.getStatusCode().isError()){
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         clientService.deleteClient(clientEntity);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
